komið

+++ include/drive.h
@@ -1,6 +1,5 @@
 #ifndef DRIVE_H
 #define DRIVE_H
-
 #include "digital_out.h"
 #include <stdint.h>
 
@@ -10,14 +9,14 @@ class Drive {
         volatile uint8_t *Pfwd;
         volatile uint8_t *Prev;
         volatile uint8_t *CTRLreg;
-        int TN;
+        uint8_t TN;
         Digital_out Pslp;
-        int speed;
+        uint16_t speed;
     public:
-        Drive(int TN_arg, int Pslp_arg);
+        Drive(uint8_t TN_arg, uint8_t Pslp_arg);
         void init(void);
-        void fwd(int speed);
-        void rev(int speed);
+        void fwd(uint16_t speed);
+        void rev(uint16_t speed);
         void stop(void); // .stop() stoppar snúning
         void sleep(void); // .sleep() stoppar snúning og slekkur á driver/brú
         void wake(void); // .wake() kveikir á driver/brú
